python-app/requirements.txt
uvicorn[standard]==0.27.0.post1
fastapi==0.109.2
websockets==12.0
pyyaml==6.0.1
aiohttp==3.9.0
aiosignal==1.2.0
anyio==4.2.0
async-timeout==4.0.3
pydantic==1.10.12
httpx==0.26.0
requests==2.31.0
openwakeword==0.6.0
tiktoken==0.6.0
openai-whisper==20231117
python-multipart==0.0.9
pyaudio==0.2.14
pygame==2.5.2
numpy==1.26.4
SpeechRecognition==3.10.3
pydub==0.25.1
pywemo==1.4.0
python-kasa==0.6.2.1

